refactor entities
